Refactor spectroscopy charts to shared helper
